abels feedback - refectoring
