+++ import.ts
@@ -1,5 +1,6 @@
 // AUTOmATICALLY GENARATED
 
 import '@/core/module/module.manager'
-import './modules/app'
-import './modules/auth'
+import './modules/auth/supabase'
+import './modules/app/recording'
+import './modules/settings'
