--- modules/app/index.ts
@@ -1,3 +1,0 @@
-import './dashboard'
-import './recording'
-import './settings'
