+++ modules/app/recording/index.ts
@@ -1,0 +1,2 @@
+import './recording'
+import './dashboard'
