+++ pages/settings/[slug].tsx
@@ -2,10 +2,11 @@ import React from "react"
 import { useRouter } from 'next/router'
 import { ModuleManager } from "@/core/module/module.manager"
 
-const App = () => {
+const Settings = () => {
     const router = useRouter()
     const { slug } = router.query as { slug: string}
-    const Component = ModuleManager.get().components.app(slug)
+    const Component = ModuleManager.get().components.settings(slug)
+
     if(!Component) {
         // TODO: Redirect to 404
         return <>Resource Not Found!</>
@@ -13,4 +14,4 @@ const App = () => {
     return <Component />
 }
 
-export default App
+export default Settings
