+++ styles/pages/register.module.css
@@ -5,8 +5,8 @@
 }
 
 .leftColumn {
-  flex: 1;
   display: flex;
+  flex: 1;
   flex-direction: column;
   justify-content: center;
   align-items: center;
@@ -20,8 +20,8 @@
 }
 
 .rightColumn {
-  flex: 1;
   display: flex;
+  flex: 1;
   justify-content: center;
   align-items: center;
   padding: 2rem;
